// src/generated/ipc/AvailableUpdate.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UpdateChannel } from "./UpdateChannel";

export type AvailableUpdate = {
  currentVersion: string;
  version: string;
  notes: string | null;
  publishedAt: string | null;
  channel: UpdateChannel;
  downloadPageUrl: string;
};